Major UI improvements
Ok so for the class I just very quickly assigned a bunch of buttons to effecting different parameters. But thats obviously confusing and would take way more work to create informative UI to explain the controls.

So instead ive opted for a simple ui where you select an option and just two buttons are for increasing and decreasing. technically 6 but they only decrease or increase. So A/B or LEFT/RIGHT or A/D