LIGHTS, CAMERA
Ok the menu is a little screwy right now but it can switch between levels which is good for me. Got the Disco Level! this is what I imagine at the end of the basement section. Manager making a office party extreme or somethin. The floors comes with 4 possible patterns you can switch in the editor. And I've got a system in place that should easily allow to do different colors