Dynamic Menus
Ok now that I'm switching between levels, Each level will have different properties to change. And I noticed when I change the visibility of a label in a Container, it automatically moves them around. So I just change the visibility depending on the level. And I just need to change the FocusNeighbors for the top and bottom of the menus.

Also added customSignlals to say when the lights should change so that I wouldn't have to update the lights colors every frame