Enemy Attack Demo
Ok got a setup that lets you preview enemy attacks and fine tune the attacks. Got 4 enemies set up. Enemy attacks can be controlled with either the progression_ratio or with the attack _pace. And then attacks can set the curve of the attackLane that they are about to go on.