Audio Manager & Most of combat
Created the audio manager and able to send signals that will create specified noises at certain events. Like when an enemy dies and is exiting the tree, or if they are damaged.

Fixed a bug where simpleCurve attacks seemed to be inconsistently blockable. Found out I was creating two instances for each attack by forgetting to remove the signal from the timer in SimpleCurve.tscn which was also creating an attack.

Basic cleanup in other spots. Ready to head back into combatFoundation to finish up this implementation.